+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## NEXT
 
   * Abort rollbacks if unexpected files are found in the /data/app/releases/ directory because they disrupt the ability to find the previous and latest releases.
+  * Accepts flags for git and archive strategies (`--git` and `--archive`) that are passed a uri
 
 ## v2.2.1 (2013-07-29)